Adapting the code to Linux 4.15
Linux 4.15 dropped setup_timer
https://github.com/torvalds/linux/commit/513ae785c63c30741e46f43960213d4ae5382ec0#diff-ba8897e9349509c7ca4d37b0704bee9c

i've tried to adapt the code using timer_setup instead, wich acts a bit differently.

The callback functions are more or less the same, and i used a global variable, "__devicepointer", to pass the pointer.

There's probably a cleaner solution than this one, but it works.